Reorganize theory module
