+++ tuprolog/theory/_adapters.py
